+++ ts/sagas/startup/loadVisibleServicesHandler.ts
@@ -1,22 +1,11 @@
-import * as pot from "italia-ts-commons/lib/pot";
 import { readableReport } from "italia-ts-commons/lib/reporters";
-import { ITuple2, Tuple2 } from "italia-ts-commons/lib/tuples";
-import { all, call, Effect, put, select } from "redux-saga/effects";
-import { PaginatedServiceTupleCollection } from "../../../definitions/backend/PaginatedServiceTupleCollection";
+import { call, Effect, put } from "redux-saga/effects";
 import { BackendClient } from "../../api/backend";
 import { sessionExpired } from "../../store/actions/authentication";
-import {
-  loadService,
-  loadVisibleServices,
-  removeServiceTuples
-} from "../../store/actions/services";
-import { messagesIdsByServiceIdSelector } from "../../store/reducers/entities/messages/messagesIdsByServiceId";
-import { servicesByIdSelector } from "../../store/reducers/entities/services/servicesById";
+import { loadVisibleServices } from "../../store/actions/services";
 import { SagaCallReturnType } from "../../types/utils";
-
-type VisibleServiceVersionById = {
-  [index: string]: number | undefined;
-};
+import { refreshStoredServices } from "../services/refreshStoredServices";
+import { removeOldStoredServices } from "../services/removeOldStoredServices";
 
 /**
  * A generator to load the service details from the Backend
@@ -53,89 +42,4 @@ export function* loadVisibleServicesRequestHandler(
   } catch (error) {
     yield put(loadVisibleServices.failure(error));
   }
-}
-
-function* removeOldStoredServices(
-  visibleServices: PaginatedServiceTupleCollection["items"]
-) {
-  const visibleServiceVersionById = visibleServices.reduce<
-    VisibleServiceVersionById
-  >(
-    (accumulator, currentValue) => ({
-      ...accumulator,
-      [currentValue.service_id]: currentValue.version
-    }),
-    {}
-  );
-
-  const storedServicesById: ReturnType<
-    typeof servicesByIdSelector
-  > = yield select(servicesByIdSelector);
-
-  const messagesIdsByServiceId: ReturnType<
-    typeof messagesIdsByServiceIdSelector
-  > = yield select(messagesIdsByServiceIdSelector);
-
-  // Create an array of tuples containing:
-  // - serviceId (to remove service from both the servicesById and the servicesMetadataById sections of the redux store)
-  // - organizationFiscalCode (to remove service from serviceIdsByOrganizationFiscalCode
-  //   section of the redux store)
-  const serviceTuplesToRemove = Object.keys(storedServicesById).reduce<
-    ReadonlyArray<ITuple2<string, string | undefined>>
-  >((accumulator, serviceId) => {
-    // Check if this service id must be removed
-    // A service must be removed if is no more visible and not used by any loaded message.
-    const mustRemoveServiceId =
-      visibleServiceVersionById[serviceId] === undefined &&
-      messagesIdsByServiceId[serviceId] === undefined;
-
-    if (!mustRemoveServiceId) {
-      return accumulator;
-    }
-
-    const storedPotService = storedServicesById[serviceId];
-
-    if (storedPotService !== undefined) {
-      // If the service detail is also loaded get the organization fiscal code
-      const organizationFiscalCode = pot.toUndefined(
-        pot.map(storedPotService, _ => _.organization_fiscal_code)
-      );
-
-      return [...accumulator, Tuple2(serviceId, organizationFiscalCode)];
-    }
-
-    return accumulator;
-  }, []);
-
-  // Dispatch action to remove the services from the redux store
-  yield put(removeServiceTuples(serviceTuplesToRemove));
-}
-
-function* refreshStoredServices(
-  visibleServices: PaginatedServiceTupleCollection["items"]
-) {
-  const storedServicesById: ReturnType<
-    typeof servicesByIdSelector
-  > = yield select(servicesByIdSelector);
-
-  const serviceDetailIdsToLoad = visibleServices
-    .filter(service => {
-      const serviceId = service.service_id;
-      const storedService = storedServicesById[serviceId];
-      return (
-        // The service detail:
-        // - is not in the redux store
-        storedService === undefined ||
-        // - is in the redux store as PotNone and not loading
-        (pot.isNone(storedService) && !pot.isLoading(storedService)) ||
-        // - is in the redux store as PotSome, is not updating and is outdated
-        (pot.isSome(storedService) &&
-          !pot.isUpdating(storedService) &&
-          storedService.value.version < service.version)
-      );
-    })
-    .map(_ => _.service_id);
-
-  // Parallel fetch of those services content that we haven't loaded yet or need to be updated
-  yield all(serviceDetailIdsToLoad.map(id => put(loadService.request(id))));
 }
